pyproject.toml
[project]
name = "kita2"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
